fix(backend): restore compilation and Spring context startup on master
Master no longer compiled (408 errors) and no backend test could run
because the Spring context failed to load. This restores the backend:

- Fix mangled annotations/imports that broke compilation across
  EventController, UserController, HackathonController/HackathonService,
  EventService, EventRepository, HackathonRepository, JwtKeyRotationManager,
  RegistrationResponse, ProjectService and AdminService.
- Restore EventRepository methods deleted by the seat-selection rewrite
  (deleteAttendeeRowsByUserId, findPublicAlternativesInWindow,
  findByTitleContainingIgnoreCaseOrDescriptionContainingIgnoreCase).
- Remove the duplicate findAlternativeEvents in EventService.
- Remove the dead decrementSeatsAtomically query that referenced the
  non-existent Event.availableSeats property and failed JPA validation
  during context startup (#13914).
- Resolve the ambiguous ObjectOptimisticLockingFailureException
  @ExceptionHandler mappings in GlobalExceptionHandler.
- Consolidate the duplicate rateLimitingFilter @Component beans into
  security/RateLimitingFilter (preserving the /api/stream/ SSE bucket
  from #13902) so the context can load.
- Align HackathonController.getAllHackathons with the Page-returning
  service contract from #14174.
- Update EventRegistrationConcurrencyIntegrationTest call sites for the
  seat-selection signature added in #8aa2c944d.

mvn clean compile and mvn test-compile now pass. The backend test suite
was previously ungated (CI only runs frontend tests), so remaining
behavioral test failures are pre-existing and out of scope here.

+++ Backend/src/main/java/com/sandeep/eventrabackend/service/ProjectService.java
@@ -127,7 +127,7 @@ public class ProjectService {
                 .category(project.getCategory())
                 .thumbnailUrl(project.getThumbnailUrl())
                 .githubUrl(project.getGithubUrl())
-                .upvotes(project.getUpvotes())
+                .upvotes((int) project.getUpvotes())
                 .build();
     }
 }
